Full review of readme.md
Changed to second person (Ie. you, your, do this) to make it more engaging.
Made the description easier to read and enumerated features as a list 
Moved the single-language feature to the language section so the organization makes more sense.
Changed the heading of the Wagtail Editor View to help content flow
Made a subtle running joke by comparing programming languages to movies.